requirements.txt
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.86.0
python-dotenv==1.0.0
sqlalchemy==2.0.19
flask==2.3.3
python-dateutil==2.8.2
requests==2.31.0
google-generativeai==0.3.2 